Did some things

+++ uu_processcontrol_maing01-server/src/main/java/uu/processcontrol/main/abl/ProcessControlAbl.java
@@ -7,18 +7,28 @@ import javax.inject.Inject;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 import uu.app.datastore.domain.PagedResult;
+import uu.app.datastore.exceptions.DatastoreIdentityRuntimeException;
+import uu.app.datastore.exceptions.DatastoreRuntimeException;
 import uu.app.exception.AppErrorMap;
 import uu.app.validation.ValidationResult;
 import uu.app.validation.Validator;
+import uu.app.validation.utils.ValidationResultUtils;
 import uu.processcontrol.main.abl.entity.Phase;
 import uu.processcontrol.main.abl.entity.PhaseCode;
 import uu.processcontrol.main.abl.entity.PhaseStatus;
 import uu.processcontrol.main.abl.entity.ProcessControl;
+import uu.processcontrol.main.abl.helper.ValidationCheckHelper;
 import uu.processcontrol.main.api.dto.ProcessControlCreateDtoIn;
+import uu.processcontrol.main.api.dto.ProcessControlCreateSingleDtoIn;
 import uu.processcontrol.main.api.dto.ProcessControlDtoOut;
 import uu.processcontrol.main.api.dto.ProcessControlGetDtoIn;
 import uu.processcontrol.main.api.dto.ProcessControlListDtoIn;
 import uu.processcontrol.main.api.dto.ProcessControlListDtoOut;
+import uu.processcontrol.main.api.dto.ProcessControlUpdateDtoIn;
+import uu.processcontrol.main.api.dto.ProcessControlValidateDtoIn;
+import uu.processcontrol.main.api.exceptions.ProcessControlRuntimeException;
+import uu.processcontrol.main.api.exceptions.ProcessControlRuntimeException.Error;
+import uu.processcontrol.main.api.utils.ProcessControlUpdateMapper;
 import uu.processcontrol.main.dao.mongo.ProcessControlMongoDao;
 
 @Component
@@ -33,47 +43,32 @@ public class ProcessControlAbl {
   @Inject
   private ProcessControlMongoDao processControlMongoDao;
 
-  public ProcessControlDtoOut create(String awid, ProcessControlCreateDtoIn dtoIn) {
+  @Inject
+  private ValidationCheckHelper validationCheckHelper;
 
+  public ProcessControlDtoOut start(String awid, ProcessControlCreateDtoIn dtoIn) {
     AppErrorMap appErrorMap = new AppErrorMap();
     //Validate dtoIn Data
     ValidationResult validationResult = validator.validate(dtoIn);
 
-    // //Checks if dtoIn valid, if no , then throws INVALID_DTO_IN exception
-    // if (!validationResult.isValid()) {
-    //   throw new TopicRuntimeException(TopicRuntimeException.Error.INVALID_DTO_IN,
-    //     ValidationResultUtils.validationResultToAppErrorMap(validationResult));
-    // }
+    if (!validationResult.isValid()) {
+      throw new ProcessControlRuntimeException(Error.INVALID_DTO_IN, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
 
-    // dtoIn = checkDefaultDate(dtoIn);
-    //Sets an object for creation
-    // Topic topic = modelMapper.map(dtoIn, Topic.class);
-    // topic.setAwid(awid);
+    try {
+      processControlMongoDao.delete(awid, dtoIn.getId());
+    } catch (DatastoreIdentityRuntimeException e) {
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_DELETE_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
 
-    //Trying to create object in database, in case of failure throws TOPIC_DAO_CREATE_FAILED exception
-    // try {
-    //   topic = topicDao.create(topic);
-    // } catch (DatastoreRuntimeException e) {
-    //   throw new TopicRuntimeException(TopicRuntimeException.Error.TOPIC_DAO_CREATE_FAILED,
-    //     ValidationResultUtils.validationResultToAppErrorMap(validationResult));
-    // }
-
-    //Sets dtoOut data for output.
-    // TopicDtoOut dtoOut = modelMapper.map(topic, TopicDtoOut.class);
-    // dtoOut.setUuAppErrorMap(appErrorMap);
-    // dtoOut.setProfiles(profiles);
-    // return dtoOut;
-
-    processControlMongoDao.delete(awid, dtoIn.getId());
-
+    //Creating empty instance for further setting
     ProcessControl processControl = new ProcessControl();
-
     processControl.setAwid(awid);
     ZonedDateTime processStart = ZonedDateTime.now();
     ZonedDateTime processEnd = processStart.plusMinutes(12);
     processControl.setStartTime(processStart);
     processControl.setEndTime(processEnd);
-
+    //Setting list of phases
     List<Phase> phases = new ArrayList<Phase>();
     //Setting receiving phase
     Phase receivingPhase = new Phase();
@@ -99,29 +94,169 @@ public class ProcessControlAbl {
     moderatingPhase.setPhaseCode(PhaseCode.Moderating);
     moderatingPhase.setStatus(PhaseStatus.INIT);
     phases.add(moderatingPhase);
-
+    //Setting phases for processControl
     processControl.setPhases(phases);
+    try {
+      processControlMongoDao.create(processControl);
+    } catch (DatastoreRuntimeException e) {
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_CREATE_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
 
-    processControlMongoDao.create(processControl);
     ProcessControlDtoOut dtoOut = modelMapper.map(processControl, ProcessControlDtoOut.class);
+    dtoOut.setUuAppErrorMap(appErrorMap);
     return dtoOut;
   }
 
-  public ProcessControlDtoOut get(String awid, ProcessControlGetDtoIn dtoIn){
+  public ProcessControlDtoOut get(String awid, ProcessControlGetDtoIn dtoIn) {
     AppErrorMap appErrorMap = new AppErrorMap();
     //Validate dtoIn Data
     ValidationResult validationResult = validator.validate(dtoIn);
 
+    if (!validationResult.isValid()) {
+      throw new ProcessControlRuntimeException(Error.INVALID_DTO_IN, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
     ProcessControl processControl = null;
-    processControl = processControlMongoDao.get(awid, dtoIn.getId());
+    try {
+      processControl = processControlMongoDao.get(awid, dtoIn.getId());
+    } catch (DatastoreIdentityRuntimeException e) {
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_GET_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
     ProcessControlDtoOut dtoOut = modelMapper.map(processControl, ProcessControlDtoOut.class);
+    dtoOut.setUuAppErrorMap(appErrorMap);
     return dtoOut;
   }
 
-  public ProcessControlListDtoOut list(String awid, ProcessControlListDtoIn dtoIn){
+  public ProcessControlListDtoOut list(String awid, ProcessControlListDtoIn dtoIn) {
+    AppErrorMap appErrorMap = new AppErrorMap();
+    //Validate dtoIn Data
+    ValidationResult validationResult = validator.validate(dtoIn);
 
-    PagedResult<ProcessControl> pagedResult = processControlMongoDao.list(awid, dtoIn.getPageInfo());
+    if (!validationResult.isValid()) {
+      throw new ProcessControlRuntimeException(Error.INVALID_DTO_IN, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    PagedResult<ProcessControl> pagedResult = null;
+
+    try {
+      pagedResult = processControlMongoDao.list(awid, dtoIn.getPageInfo());
+    } catch (DatastoreIdentityRuntimeException e) {
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_LIST_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
     ProcessControlListDtoOut dtoOut = modelMapper.map(pagedResult, ProcessControlListDtoOut.class);
+    dtoOut.setUuAppErrorMap(appErrorMap);
+    return dtoOut;
+  }
+
+  public ProcessControlDtoOut createSingle(String awid, ProcessControlCreateSingleDtoIn dtoIn){
+    AppErrorMap appErrorMap = new AppErrorMap();
+    //Validate dtoIn Data
+    ValidationResult validationResult = validator.validate(dtoIn);
+
+    if (!validationResult.isValid()) {
+      throw new ProcessControlRuntimeException(Error.INVALID_DTO_IN, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    //Creating empty instance for further setting
+    ProcessControl processControl = new ProcessControl();
+    processControl.setAwid(awid);
+    ZonedDateTime processStart = ZonedDateTime.now();
+    ZonedDateTime processEnd = processStart.plusMinutes(12);
+    processControl.setStartTime(processStart);
+    processControl.setEndTime(processEnd);
+    //Setting list of phases
+    List<Phase> phases = new ArrayList<Phase>();
+    //Setting receiving phase
+    Phase receivingPhase = new Phase();
+    receivingPhase.setStartTime(processStart);
+    receivingPhase.setEndTime(processStart.plusMinutes(5));
+    receivingPhase.setName("Receiving");
+    receivingPhase.setPhaseCode(PhaseCode.Receiving);
+    receivingPhase.setStatus(PhaseStatus.RUNNING);
+    phases.add(receivingPhase);
+    //Setting validationPhase
+    Phase validationPhase = new Phase();
+    validationPhase.setStartTime(receivingPhase.getEndTime());
+    validationPhase.setEndTime(validationPhase.getStartTime().plusMinutes(2));
+    validationPhase.setName("Validation");
+    validationPhase.setPhaseCode(PhaseCode.Validation);
+    validationPhase.setStatus(PhaseStatus.INIT);
+    phases.add(validationPhase);
+    //Setting moderatingPhase
+    Phase moderatingPhase = new Phase();
+    moderatingPhase.setStartTime(validationPhase.getEndTime());
+    moderatingPhase.setEndTime(moderatingPhase.getStartTime().plusMinutes(5));
+    moderatingPhase.setName("Moderating");
+    moderatingPhase.setPhaseCode(PhaseCode.Moderating);
+    moderatingPhase.setStatus(PhaseStatus.INIT);
+    phases.add(moderatingPhase);
+    //Setting phases for processControl
+    processControl.setPhases(phases);
+    try {
+      processControlMongoDao.create(processControl);
+    } catch (DatastoreRuntimeException e) {
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_CREATE_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    ProcessControlDtoOut dtoOut = modelMapper.map(processControl, ProcessControlDtoOut.class);
+    dtoOut.setUuAppErrorMap(appErrorMap);
+    return dtoOut;
+  }
+
+  public ProcessControlDtoOut update(String awid, ProcessControlUpdateDtoIn dtoIn){
+    AppErrorMap appErrorMap = new AppErrorMap();
+    //Validate dtoIn Data
+    ValidationResult validationResult = validator.validate(dtoIn);
+
+    if (!validationResult.isValid()) {
+      throw new ProcessControlRuntimeException(Error.INVALID_DTO_IN, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    ProcessControl processControl = new ProcessControl();
+    try {
+      processControl = processControlMongoDao.get(awid, dtoIn.getId());
+    } catch (DatastoreIdentityRuntimeException e) {
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_GET_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    if(processControl != null){
+      ProcessControlUpdateMapper updateMapper = new ProcessControlUpdateMapper();
+      processControl = updateMapper.map(processControl, dtoIn);
+      processControl = processControlMongoDao.update(processControl);
+    }else{
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_UPDATE_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    ProcessControlDtoOut dtoOut = modelMapper.map(processControl, ProcessControlDtoOut.class);
+    dtoOut.setUuAppErrorMap(appErrorMap);
+    return dtoOut;
+  }
+
+  public ProcessControlDtoOut validate(String awid, ProcessControlValidateDtoIn dtoIn){
+    AppErrorMap appErrorMap = new AppErrorMap();
+    //Validate dtoIn Data
+    ValidationResult validationResult = validator.validate(dtoIn);
+
+    if (!validationResult.isValid()) {
+      throw new ProcessControlRuntimeException(Error.INVALID_DTO_IN, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    ProcessControl processControl = new ProcessControl();
+    try {
+      processControl = processControlMongoDao.get(awid, dtoIn.getId());
+    } catch (DatastoreIdentityRuntimeException e) {
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_GET_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+
+    if(processControl != null){
+      processControl.getPhases().set(1, validationCheckHelper.checkValidationPhase(processControl));
+      processControl = processControlMongoDao.update(processControl);
+    }else{
+      throw new ProcessControlRuntimeException(Error.PROCESS_DAO_UPDATE_FAILED, ValidationResultUtils.validationResultToAppErrorMap(validationResult));
+    }
+    ProcessControlDtoOut dtoOut = modelMapper.map(processControl, ProcessControlDtoOut.class);
+    dtoOut.setUuAppErrorMap(appErrorMap);
     return dtoOut;
   }
 }
